+++ symbolic-debuginfo/src/pe.rs
@@ -154,8 +154,12 @@ impl<'data> PeObject<'data> {
             .debug_data
             .as_ref()
             .and_then(|debug_data| debug_data.codeview_pdb70_debug_info.as_ref())
-            .map(|debug_info| {
-                String::from_utf8_lossy(&debug_info.filename[..debug_info.filename.len() - 1])
+            .and_then(|debug_info| {
+                debug_info
+                    .filename
+                    .iter()
+                    .position(|&c| c == 0)
+                    .map(|nul_byte| String::from_utf8_lossy(&debug_info.filename[..nul_byte]))
             })
     }